requirements.txt
asttokens==2.4.1
chardet==5.2.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.2
decorator==5.1.1
executing==2.0.1
greenlet==3.0.3
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2==2.9.9
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyzmq==26.0.3
six==1.16.0
SQLAlchemy==2.0.31
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
wcwidth==0.2.13
